use renaming function to rename raw photos
In this case of renaming the original raw photos, some parts of the
function are redundant. For example, when checking if the original file
exists. This is valid for XMP and JPG files because the code checks if
one of these files were derived from the original. But when renaming the
original raw file, it has been determined that the file to be renamed
exists. Still, I determined it to be better to repeat this small part of
the process in order to make the code look cleaner.